show webpage on detail pane
