+++ app/src/main/java/liou/rayyuan/ebooksearchtaiwan/di/AppModule.kt
@@ -15,7 +15,6 @@ import liou.rayyuan.ebooksearchtaiwan.interactor.UserRankingWindowFacade
 import liou.rayyuan.ebooksearchtaiwan.model.DeeplinkHelper
 import liou.rayyuan.ebooksearchtaiwan.model.EventTracker
 import liou.rayyuan.ebooksearchtaiwan.preferencesetting.PreferenceSettingsViewModel
-import liou.rayyuan.ebooksearchtaiwan.simplewebview.SimpleWebViewViewModel
 import liou.rayyuan.ebooksearchtaiwan.utils.ClipboardHelper
 import liou.rayyuan.ebooksearchtaiwan.utils.CustomTabSessionManager
 import liou.rayyuan.ebooksearchtaiwan.utils.DeviceVibrateHelper
@@ -64,10 +63,6 @@ val appModule =
             PreferenceSettingsViewModel(
                 deleteAllSearchRecord = get()
             )
-        }
-
-        viewModel {
-            SimpleWebViewViewModel(get())
         }
 
         // Interactors
